Adopt idiomatic C-style for error returns (#132)
* Safely call free on zenoh types

* Un-const links in transports

* First pass on check for NULL upon allocation

* _z_string_from_bytes accepts a const _z_bytes_t

* Leverage on values for internal session instead of pointers

* Return int8_t from system specific functions

* Removal of _z_*_result_t types

* Reordering members in structs

* Improve decoding readability

* Organize error codes

* Fix unsafe clear during init-open procedure in case of error

* Mode configuration error

* Extra validations on the endpoint format based underlying protocol

* Group IF statements on INIT-OPEN

* Properly clean session and transport upon err

* Fix set error on mbed endpoint creation

* Reintroduce test values

* Remove err state from link-related procedures

* Add missing message copy helpers

* Make sure that messages are copy and clear safe no matter the flags

* Do not consider the 0x00 delimeter on the decobs size

* Also check for the serial support to auto-disable unicast transport

* Fix multicast transport initialization

* Add missing consts on zephyr

* Bluetooth gname is an alias to endpoint address

* Address PR review comments

* Avoid double free on z_drop

* Get Bluetooth timeout from locator config

* Fix task join/cancel in arduino-esp32

* Add missing consts in system network

* Continue if ret is still OK

* Fix task join/cancel in espidf

* Fix network clean in opencr port

* Unlock in case of error

* Handle return as int8_t

* Fix function signature

* Change session initialization due to issue on zephyr

* Final MISRA pass on the PR

* Smart pointer counter allocataion iff value is allocated

* Rename Peer ID to Zenoh ID

* Make ZID 16 bytes by default

* Fix keyexpr-to-string release in examples

* Small fix on example

+++ include/zenoh-pico/system/platform.h
@@ -50,25 +50,25 @@ void z_free(void *ptr);
 
 #if Z_MULTI_THREAD == 1
 /*------------------ Thread ------------------*/
-int _z_task_init(_z_task_t *task, _z_task_attr_t *attr, void *(*fun)(void *), void *arg);
-int _z_task_join(_z_task_t *task);
-int _z_task_cancel(_z_task_t *task);
+int8_t _z_task_init(_z_task_t *task, _z_task_attr_t *attr, void *(*fun)(void *), void *arg);
+int8_t _z_task_join(_z_task_t *task);
+int8_t _z_task_cancel(_z_task_t *task);
 void _z_task_free(_z_task_t **task);
 
 /*------------------ Mutex ------------------*/
-int _z_mutex_init(_z_mutex_t *m);
-int _z_mutex_free(_z_mutex_t *m);
+int8_t _z_mutex_init(_z_mutex_t *m);
+int8_t _z_mutex_free(_z_mutex_t *m);
 
-int _z_mutex_lock(_z_mutex_t *m);
-int _z_mutex_trylock(_z_mutex_t *m);
-int _z_mutex_unlock(_z_mutex_t *m);
+int8_t _z_mutex_lock(_z_mutex_t *m);
+int8_t _z_mutex_trylock(_z_mutex_t *m);
+int8_t _z_mutex_unlock(_z_mutex_t *m);
 
 /*------------------ CondVar ------------------*/
-int _z_condvar_init(_z_condvar_t *cv);
-int _z_condvar_free(_z_condvar_t *cv);
+int8_t _z_condvar_init(_z_condvar_t *cv);
+int8_t _z_condvar_free(_z_condvar_t *cv);
 
-int _z_condvar_signal(_z_condvar_t *cv);
-int _z_condvar_wait(_z_condvar_t *cv, _z_mutex_t *m);
+int8_t _z_condvar_signal(_z_condvar_t *cv);
+int8_t _z_condvar_wait(_z_condvar_t *cv, _z_mutex_t *m);
 #endif  // Z_MULTI_THREAD == 1
 
 /*------------------ Sleep ------------------*/
