+++ src/system/unix/system.c
@@ -98,13 +98,13 @@ void z_free(void *ptr) { free(ptr); }
 
 #if Z_MULTI_THREAD == 1
 /*------------------ Task ------------------*/
-int _z_task_init(_z_task_t *task, pthread_attr_t *attr, void *(*fun)(void *), void *arg) {
+int8_t _z_task_init(_z_task_t *task, pthread_attr_t *attr, void *(*fun)(void *), void *arg) {
     return pthread_create(task, attr, fun, arg);
 }
 
-int _z_task_join(_z_task_t *task) { return pthread_join(*task, NULL); }
+int8_t _z_task_join(_z_task_t *task) { return pthread_join(*task, NULL); }
 
-int _z_task_cancel(_z_task_t *task) { return pthread_cancel(*task); }
+int8_t _z_task_cancel(_z_task_t *task) { return pthread_cancel(*task); }
 
 void _z_task_free(_z_task_t **task) {
     _z_task_t *ptr = *task;
@@ -113,24 +113,24 @@ void _z_task_free(_z_task_t **task) {
 }
 
 /*------------------ Mutex ------------------*/
-int _z_mutex_init(_z_mutex_t *m) { return pthread_mutex_init(m, 0); }
+int8_t _z_mutex_init(_z_mutex_t *m) { return pthread_mutex_init(m, 0); }
 
-int _z_mutex_free(_z_mutex_t *m) { return pthread_mutex_destroy(m); }
+int8_t _z_mutex_free(_z_mutex_t *m) { return pthread_mutex_destroy(m); }
 
-int _z_mutex_lock(_z_mutex_t *m) { return pthread_mutex_lock(m); }
+int8_t _z_mutex_lock(_z_mutex_t *m) { return pthread_mutex_lock(m); }
 
-int _z_mutex_trylock(_z_mutex_t *m) { return pthread_mutex_trylock(m); }
+int8_t _z_mutex_trylock(_z_mutex_t *m) { return pthread_mutex_trylock(m); }
 
-int _z_mutex_unlock(_z_mutex_t *m) { return pthread_mutex_unlock(m); }
+int8_t _z_mutex_unlock(_z_mutex_t *m) { return pthread_mutex_unlock(m); }
 
 /*------------------ Condvar ------------------*/
-int _z_condvar_init(_z_condvar_t *cv) { return pthread_cond_init(cv, 0); }
+int8_t _z_condvar_init(_z_condvar_t *cv) { return pthread_cond_init(cv, 0); }
 
-int _z_condvar_free(_z_condvar_t *cv) { return pthread_cond_destroy(cv); }
+int8_t _z_condvar_free(_z_condvar_t *cv) { return pthread_cond_destroy(cv); }
 
-int _z_condvar_signal(_z_condvar_t *cv) { return pthread_cond_signal(cv); }
+int8_t _z_condvar_signal(_z_condvar_t *cv) { return pthread_cond_signal(cv); }
 
-int _z_condvar_wait(_z_condvar_t *cv, _z_mutex_t *m) { return pthread_cond_wait(cv, m); }
+int8_t _z_condvar_wait(_z_condvar_t *cv, _z_mutex_t *m) { return pthread_cond_wait(cv, m); }
 #endif  // Z_MULTI_THREAD == 1
 
 /*------------------ Sleep ------------------*/
